requirements.txt
Django==5.0.6
djangorestframework==3.15.2
Pillow==11.3.0
django-cors-headers==4.3.1
